+++ frontend/src/index.js
@@ -1,6 +1,34 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux';
+import reducer from "./reducers";
+import thunk from 'redux-thunk';
 import App from './Components/App';
+import { checkLoginUser } from "./actions";
 import "./api";
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const store = createStore(
+  reducer,
+  applyMiddleware(thunk)
+);
+
+async function check() {
+  //login 유효 검사
+  const jwtToken = localStorage.getItem("jwtToken");
+  try {
+    if(jwtToken) {
+      await store.dispatch(checkLoginUser(jwtToken));
+    }
+  } catch(e) {
+    console.log('error', e);
+  }
+};
+
+check();
+
+ReactDOM.render(
+  <Provider store={store}>
+    <App />
+  </Provider>, document.getElementById('root')
+);
